Splitting up model.py to several files.
